+++ backend/src/shared/websockets/poker-tables/types/PokerTableEvents.ts
@@ -7,3 +7,8 @@ export type playerLeftEvent = {
   username: string;
   seatNumber: string;
 };
+
+export interface PokerTableEvents {
+  player_joined: (payload: playerJoinedEvent) => void;
+  player_left: (payload: playerLeftEvent) => void;
+}
