+++ backend/src/sockets/SocketEventHandlers.ts
@@ -1,6 +1,6 @@
 import { Socket } from 'socket.io';
 
-import { ClientToServerEvents, ServerToClientEvents } from '@shared/websockets/SocketEvents';
+import { ClientToServerEvents, ServerToClientEvents } from '@shared/websockets/WebsocketEvents';
 
 import { Logger } from '../utils/Logger';
